Revise per @laurejt 's review
