+++ pyproject.toml
@@ -28,6 +28,7 @@ dependencies = [
   "numpy",
   "ipython", # Required by transformers for trainer functionality
   "transformers[torch, sentencepiece, tiktoken]",
+  "orjsonl>=1.0.0",
 ]
 
 [project.optional-dependencies]
