--- src/muse/scripts/__init__.py
@@ -1,5 +1,0 @@
-"""Scripts for machine translation corpus generation."""
-
-from muse.scripts.generate_mt_corpus import main
-
-__all__ = ["main"]
